requirements.txt
# HubSpot Integration
requests>=2.28.0
python-dotenv>=1.0.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Email (optional)
sendgrid>=6.10.0

# PDF Generation (for certificates)
reportlab>=4.0.0
